Add Introduction container, caption for img, Title|
The page wasn't looking so organized, so I setup a title page with its
own div box entitled "MY WONDERFUL EXPERIENCE AT SONOVA", and then
included a panel for the intro. The intro panel is flexed with two
containers inside: intro .textBox and intro .introImage.

I designed this part to look like the format of a wikipedia page because
I think it looks simple and neat. I will make the rest of the containers
look like this. Also, the sidebar links are now working as intended.